requirements.txt
autopep8==1.3.4
click==6.7
configparser==3.5.0
enum34==1.1.6
flake8==3.5.0
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
py2neo==3.1.2
pycodestyle==2.3.1
pyflakes==1.6.0
PyVCF==0.6.8
tqdm==4.19.6
